backend/bot/requirements.txt
# Core Discord bot framework
discord.py==2.3.2
aiohttp==3.9.1

# Database integration
psycopg2-binary==2.9.9
SQLAlchemy==2.0.23
alembic==1.12.1
greenlet==3.0.1

# Async database support
asyncpg==0.29.0

# Task scheduling
APScheduler==3.10.4
APScheduler[postgres]==3.10.4

# HTTP client for external APIs
httpx==0.25.2
requests==2.31.0
aiohttp-socks==0.8.4

# Media API integrations
tmdbv3api==1.8.3
gql[all]==3.4.1
graphql-core==3.2.3

# LLM integration
openai==1.3.7

# Configuration management
python-decouple==3.8
pydantic==2.5.0
python-dotenv==1.0.0

# Logging and monitoring
structlog==23.2.0
python-json-logger==2.0.7
loguru==0.7.2

# Date/time utilities
arrow==1.2.3
python-dateutil==2.8.2
pytz==2023.3

# Security utilities
cryptography==41.0.7
PyJWT==2.8.0
passlib[bcrypt]==1.7.4

# Data processing
orjson==3.9.10
pandas==2.1.3

# Image processing for embeds
Pillow==10.1.0

# Async utilities
asyncio-mqtt==0.16.0
aiofiles==23.2.1

# Background task processing
celery==5.3.4
redis==5.0.1

# Validation and serialization
pydantic-settings==2.1.0
marshmallow==3.20.1

# Discord interactions
discord-py-slash-command==3.0.1

# External API clients
python-tmdbv3api==1.8.3

# Caching
redis==5.0.1
aioredis==2.0.1

# Metrics and monitoring
prometheus-client==0.19.0

# Testing and development
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-cov==4.1.0
black==23.10.1
isort==5.12.0
flake8==6.1.0
mypy==1.6.1
bandit==1.10.1

# Type hints
types-requests==2.31.0.20231115
types-PyJWT==2.8.0.20231013
types-python-dateutil==2.8.19.20230718
typing-extensions==4.8.0

# Production deployment
gunicorn==21.2.0

# Utilities
tenacity==8.2.3
python-multipart==0.0.6

# Version pinning for production
# Note: Pin exact versions after testing in production environment